feat(oa4rust): 多 crate 控制层与组织层实现更新
- auth/person: 人员领域模型完善
- control/{group,person,role,unit}: 控制层 CRUD 与权限细化
- empower/src/lib: 授权能力增强
- file/general/meeting/message_assemble_control: assembler 控制逻辑更新
- organization_assemble_control: 组织/人员/ misc 查询实现
- portal_assemble_{designer,surface}: 门户设计器与表面层
- program_center: 程序中心实现
- query_assemble_{designer,surface,core_express}: 查询组装与表达式核心

+++ oa4rust/crates/control/src/group.rs
@@ -93,7 +93,7 @@ async fn query_page(
             .query(
                 "SELECT id, name, description, disable FROM auth_group \
                  WHERE deleted_at IS NULL AND (name > $1 OR $1 = '' OR $1 = '-') \
-                 ORDER BY name ASC LIMIT $2::int",
+                 ORDER BY name ASC LIMIT $2",
                 &[&flag, &count],
             )
             .await
@@ -103,7 +103,7 @@ async fn query_page(
             .query(
                 "SELECT id, name, description, disable FROM auth_group \
                  WHERE deleted_at IS NULL AND (name < $1 OR $1 = '' OR $1 = '-') \
-                 ORDER BY name DESC LIMIT $2::int",
+                 ORDER BY name DESC LIMIT $2",
                 &[&flag, &count],
             )
             .await
