+++ oa4rust/crates/control/src/person.rs
@@ -142,7 +142,7 @@ async fn query_page(
             .query(
                 "SELECT id, unique_id, name, mobile, email, locked FROM auth_person \
                  WHERE deleted_at IS NULL AND (unique_id > $1 OR $1 = '' OR $1 = '-') \
-                 ORDER BY unique_id ASC LIMIT $2::int",
+                 ORDER BY unique_id ASC LIMIT $2",
                 &[&flag, &limit],
             )
             .await
@@ -154,7 +154,7 @@ async fn query_page(
             .query(
                 "SELECT id, unique_id, name, mobile, email, locked FROM auth_person \
                  WHERE deleted_at IS NULL AND (unique_id < $1 OR $1 = '' OR $1 = '-') \
-                 ORDER BY unique_id DESC LIMIT $2::int",
+                 ORDER BY unique_id DESC LIMIT $2",
                 &[&flag, &limit],
             )
             .await
