+++ oa4rust/crates/control/src/unit.rs
@@ -150,7 +150,7 @@ async fn query_page(
             .query(
                 "SELECT id, name, parent_id, level FROM auth_unit \
                  WHERE deleted_at IS NULL AND (name > $1 OR $1 = '' OR $1 = '-') \
-                 ORDER BY name ASC LIMIT $2::int",
+                 ORDER BY name ASC LIMIT $2",
                 &[&flag, &count],
             )
             .await
@@ -160,7 +160,7 @@ async fn query_page(
             .query(
                 "SELECT id, name, parent_id, level FROM auth_unit \
                  WHERE deleted_at IS NULL AND (name < $1 OR $1 = '' OR $1 = '-') \
-                 ORDER BY name DESC LIMIT $2::int",
+                 ORDER BY name DESC LIMIT $2",
                 &[&flag, &count],
             )
             .await
